+++ src/Integration.php
@@ -7,7 +7,7 @@
  * Company: Pronamic
  *
  * @author Remco Tolsma
- * @version 1.1.5
+ * @version 1.1.6
  * @since 1.0.0
  */
 class Pronamic_WP_Pay_Gateways_Mollie_Integration extends Pronamic_WP_Pay_Gateways_AbstractIntegration {
@@ -30,6 +30,11 @@ class Pronamic_WP_Pay_Gateways_Mollie_Integration extends Pronamic_WP_Pay_Gatewa
 
 		if ( ! has_action( 'wp_loaded', $function ) ) {
 			add_action( 'wp_loaded', $function );
+		}
+
+		if ( is_admin() ) {
+			add_action( 'show_user_profile', array( $this, 'user_profile' ) );
+			add_action( 'edit_user_profile', array( $this, 'user_profile' ) );
 		}
 	}
 
@@ -54,5 +59,15 @@ class Pronamic_WP_Pay_Gateways_Mollie_Integration extends Pronamic_WP_Pay_Gatewa
 		$settings[] = 'mollie';
 
 		return $settings;
+	}
+
+	/**
+	 * User profile.
+	 *
+	 * @since 1.1.6
+	 * @see https://github.com/WordPress/WordPress/blob/4.5.2/wp-admin/user-edit.php#L578-L600
+	 */
+	public function user_profile( $user ) {
+		include dirname( __FILE__ ) . '/../views/html-admin-user-profile.php';
 	}
 }