+++ CHANGELOG.md
@@ -7,6 +7,13 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [1.1.6] - 2016-06-08
+- Added support for Mollie Checkout.
+- Reduced the use of else expressions.
+- Added WordPress payment method to Mollie method transform function.
+- Added readonly Mollie user profile fields.
+- Simplified the gateay payment start function.
+
 ## [1.1.5] - 2016-03-22
 - Added product URL, updated dashboard URL.
 - Updated gateway settings.
@@ -35,7 +42,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2015-01-19
 - First release.
 
-[unreleased]: https://github.com/wp-pay-gateways/mollie/compare/1.1.5...HEAD
+[unreleased]: https://github.com/wp-pay-gateways/mollie/compare/1.1.6...HEAD
+[1.1.6]: https://github.com/wp-pay-gateways/mollie/compare/1.1.5...1.1.6
 [1.1.5]: https://github.com/wp-pay-gateways/mollie/compare/1.1.4...1.1.5
 [1.1.4]: https://github.com/wp-pay-gateways/mollie/compare/1.1.3...1.1.4
 [1.1.3]: https://github.com/wp-pay-gateways/mollie/compare/1.1.2...1.1.3